requirements.txt
accessible-output2==0.17
altgraph==0.17.4
babel==2.17.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
defusedxml==0.7.1
fonttools==4.59.2
fpdf2==2.8.4
kiwisolver==1.4.9
libloader==1.3.3
lxml==6.0.1
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
platform_utils==1.5.10
platformdirs==4.4.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
setuptools==80.9.0
six==1.17.0
tkcalendar==1.6.1
typing_extensions==4.15.0

# To install the dependencies:
# pip install -r requirements.txt

# To prevent creation of .pyc files:
# [Environment]::SetEnvironmentVariable("PYTHONDONTWRITEBYTECODE", "1", "User")